bankinterface :loan_customer 房屋面积 字段类型修改

+++ src/main/java/com/yunche/loan/domain/param/AppCustomerParam.java
@@ -119,7 +119,7 @@ public class AppCustomerParam {
 
     private Long feedingNum;
 
-    private Long floorSpace;
+    private BigDecimal floorSpace;
 
     private Byte houseOwnerRelation;
